More care in the use of long. Comments about what will probably be needed on Win64.
git-svn-id: https://svn.r-project.org/R/trunk@35334 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/src/carray.h
@@ -29,7 +29,7 @@ Array make_identity_matrix(int n);
 
 Array subarray(Array a, int index);
 
-long vector_length(Array a);
+int vector_length(Array a);
 
 void set_array_to_zero(Array arr);
 void copy_array (Array orig, Array ans);
